+++ tasks/seq/kholin_k_multidimensional_integrals_rectangle/include/ops_seq.hpp
@@ -27,20 +27,20 @@ class TestTaskSequential : public ppc::core::Task {
   double start_n_;
   double result_;
 
-  double dim_;
+  size_t dim_;
   size_t sz_values_;
   size_t sz_lower_limits_;
   size_t sz_upper_limits_;
 
   double Integrate(const Function& f, const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-                   const std::vector<double>& h, std::vector<double>& f_values, int curr_index_dim, double dim,
+                   const std::vector<double>& h, std::vector<double>& f_values, int curr_index_dim, size_t dim,
                    double n);
   double IntegrateWithRectangleMethod(const Function& f, std::vector<double>& f_values,
                                       const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-                                      double dim, double n);
+                                      size_t dim, double n);
   double RunMultistepSchemeMethodRectangle(double epsilon, const Function& f, std::vector<double>& f_values,
                                            const std::vector<double>& l_limits, const std::vector<double>& u_limits,
-                                           double dim, double n);
+                                           size_t dim, double n);
 };
 
 }  // namespace kholin_k_multidimensional_integrals_rectangle_seq
